claim_audit_xai_public_v2/requirements.txt
requests
beautifulsoup4
pdfplumber
faiss-cpu
openai
playwright
llama-cloud-services
